i32 wasn't enough anymore :)

+++ advent-of-code-2024/src/day1.rs
@@ -7,19 +7,19 @@ pub struct Day1 {
 }
 
 impl Day1 {
-    fn get_numbers(line: &str) -> (i32, i32) {
+    fn get_numbers(line: &str) -> (i128, i128) {
         let nums = line
             .split_whitespace()
-            .map(|x| x.parse::<i32>().unwrap())
-            .collect::<Vec<i32>>();
+            .map(|x| x.parse::<i128>().unwrap())
+            .collect::<Vec<i128>>();
         (nums[0], nums[1])
     }
 }
 
 impl Solution for Day1 {
-    fn part1(&self) -> i32 {
-        let mut first_set: Vec<i32> = Vec::new();
-        let mut second_set: Vec<i32> = Vec::new();
+    fn part1(&self) -> i128 {
+        let mut first_set: Vec<i128> = Vec::new();
+        let mut second_set: Vec<i128> = Vec::new();
 
         for line in self.input.split("\n") {
             let nums = Self::get_numbers(line);
@@ -39,9 +39,9 @@ impl Solution for Day1 {
         sum.try_into().unwrap()
     }
 
-    fn part2(&self) -> i32 {
-        let mut first_set: Vec<i32> = Vec::new();
-        let mut second_set: Vec<i32> = Vec::new();
+    fn part2(&self) -> i128 {
+        let mut first_set: Vec<i128> = Vec::new();
+        let mut second_set: Vec<i128> = Vec::new();
 
         for line in self.input.split("\n") {
             let nums = Self::get_numbers(line);
@@ -50,7 +50,7 @@ impl Solution for Day1 {
             second_set.push(nums.1);
         }
 
-        let mut frequency_map: HashMap<i32, i32> = HashMap::new();
+        let mut frequency_map: HashMap<i128, i128> = HashMap::new();
 
         for num in second_set {
             *frequency_map.entry(num).or_insert(0) += 1;
@@ -59,7 +59,7 @@ impl Solution for Day1 {
         let mut sum = 0;
 
         for num in first_set {
-            sum += num * frequency_map.get(&num).unwrap_or(&0i32);
+            sum += num * frequency_map.get(&num).unwrap_or(&0i128);
         }
 
         sum
